feature(findByQuery): middleware for find by query like pagination, username, etc...
